move bts logic to paymentComponent
