+++ src/components/GiantArtCard/index.tsx
@@ -6,6 +6,7 @@ import AuthorComponent from 'pages/Home/HotAuction/components/AuthorComponent';
 import DescriptionAndTagsComponent from 'pages/Home/HotAuction/components/DescriptionAndTagsComponent';
 import PaymentComponent from 'pages/Home/HotAuction/components/PaymentComponent';
 import ViewsAndControlsComponent from 'pages/Home/HotAuction/components/ViewsAndControlsComponent';
+import { INft } from '../../typings';
 
 import styles from './styles.module.scss';
 
@@ -18,73 +19,41 @@ type Props = {
   link: string;
   likeAction: () => void;
   dotsAction: () => void;
-  price: number | string;
-  asset: string;
   growth: number;
   growthUsd: number;
-  author: string;
-  authorAvatar: string;
-  tags: string[];
-  description: string;
-  image: string;
-  isUserCanBuyNft: boolean;
-  isUserCanEnterInAuction: boolean;
-  isUserCanPutOnSale: boolean;
-  type: 'auction' | 'sell' | '';
+  nft: INft | null;
 };
 
 const GiantCard: FC<Props> = ({
   className,
-  name,
-  likes,
   views,
-  inStock,
   link,
   likeAction,
   dotsAction,
-  price,
-  asset,
   growth,
   growthUsd,
-  author,
-  authorAvatar,
-  tags,
-  description,
-  image,
-  isUserCanBuyNft,
-  isUserCanEnterInAuction,
-  isUserCanPutOnSale,
-  type,
+  nft,
 }) => (
   <div className={cx(styles.giantCard, className)}>
-    <img src={image} alt="" />
+    <img src={nft?.media || ''} alt="" />
     <div className={styles.cardInfo}>
-      <H2>{name}</H2>
+      <H2>{nft?.name || ''}</H2>
       <ViewsAndControlsComponent
         className={styles.detailedViewsAndControl}
-        likes={likes}
+        likes={nft?.like_count || 0}
         views={views}
-        inStock={inStock}
+        inStock={nft?.available || 0}
         link={link}
         likeAction={likeAction}
         dotsAction={dotsAction}
       />
-      {type ? (
-        <PaymentComponent
-          price={price.toString(10)}
-          asset={asset}
-          type={type}
-          growthUsd={growthUsd}
-          growth={growth}
-          isUserCanBuyNft={isUserCanBuyNft}
-          isUserCanEnterInAuction={isUserCanEnterInAuction}
-          isUserCanPutOnSale={isUserCanPutOnSale}
-        />
+      {nft?.is_auc_selling || nft?.is_selling ? (
+        <PaymentComponent growthUsd={growthUsd} growth={growth} nft={nft} />
       ) : (
         ''
       )}
-      <AuthorComponent author={author} authorPic={authorAvatar} />
-      <DescriptionAndTagsComponent tags={tags} body={description} />
+      <AuthorComponent author={nft?.creator.name || ''} authorPic={nft?.creator.avatar || ''} />
+      <DescriptionAndTagsComponent tags={nft?.tags || []} body={nft?.description || ''} />
     </div>
   </div>
 );
