+++ src/pages/Detail/Artwork.tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react';
 import cx from 'classnames';
 import { useParams, useHistory } from 'react-router-dom';
-import { observer } from 'mobx-react-lite';
 
 import { ArtCard, Button, GiantCard, H3, Select, Text, TradingHistory, Control } from 'components';
 import {
@@ -10,9 +9,8 @@ import {
   TradingHistoryPrice,
 } from 'components/Table/TradingHistoryCells';
 import { Chart } from 'containers';
-import { TableCell, INft, IOwner } from 'typings';
+import { TableCell, INft } from 'typings';
 import { storeApi } from '../../services/api';
-import { useMst } from '../../store';
 
 import { artworkData, data as mockData, tableDataArtwork } from './mockdata';
 
@@ -89,136 +87,14 @@ const columnTest = [
   },
 ];
 
-const DetailArtwork: FC<Props> = observer(({ className }) => {
+const DetailArtwork: FC<Props> = ({ className }) => {
   const history = useHistory();
 
   const { id } = useParams<{ id: string }>();
 
-  const { user } = useMst();
-
   const [nft, setNft] = React.useState<INft | null>(null);
 
   console.log('nft data', nft);
-
-  const isOwner = React.useMemo(() => {
-    if (nft && nft.owners) {
-      if (Array.isArray(nft.owners)) {
-        return !!nft.owners.find((owner: IOwner) => {
-          return owner.id === user.id;
-        });
-      }
-      return user.id === nft.owners.id;
-    }
-    return false;
-  }, [nft, user.id]);
-
-  console.log('isOwner', isOwner);
-
-  const isWrongChain = React.useMemo(() => {
-    if (!nft) return false;
-    if (
-      nft?.network.name === 'Binance-Smart-Chain' &&
-      localStorage.netfly_nft_chainName === 'Binance'
-    ) {
-      return false;
-    }
-    if (
-      nft?.network.name === 'Kardiachain' &&
-      localStorage.netfly_nft_chainName === 'KardiaChain'
-    ) {
-      return false;
-    }
-    return true;
-  }, [nft]);
-
-  console.log('isWrongChain', isWrongChain);
-
-  const isUserCanRemoveFromSale = React.useMemo(() => {
-    if (nft && !isWrongChain) {
-      if (nft.standart === 'ERC721' && nft.is_selling && isOwner) return true;
-      if (
-        nft.standart === 'ERC1155' &&
-        (nft.sellers.find((seller) => seller.id === user.id) ||
-          nft.owner_auction.find((seller) => seller.id === user.id)) &&
-        isOwner
-      )
-        return true;
-    }
-    return false;
-  }, [nft, isOwner, user.id, isWrongChain]);
-
-  console.log('isUserCanRemoveFromSale', isUserCanRemoveFromSale);
-
-  const isUserCanBuyNft = React.useMemo(() => {
-    if (nft && !isWrongChain && nft.price && nft.is_selling && nft.available !== 0) {
-      if (nft.standart === 'ERC721' && !isOwner) return true;
-      if (
-        nft.standart === 'ERC1155' &&
-        ((nft.sellers.length === 1 && nft.sellers[0].id !== user.id) || nft.sellers.length > 1)
-      )
-        return true;
-    }
-    return false;
-  }, [nft, user.id, isOwner, isWrongChain]);
-
-  console.log('isUserCanBuyNft', isUserCanBuyNft);
-
-  const isUserCanEnterInAuction = React.useMemo(() => {
-    if (nft && !isWrongChain && nft.is_auc_selling && nft.available !== 0) {
-      if (nft.standart === 'ERC721' && !isOwner) return true;
-      if (
-        nft.standart === 'ERC1155' &&
-        (nft.owner_auction.length > 1 ||
-          (nft.owner_auction.length === 1 && nft.owner_auction[0].id !== user.id))
-      )
-        return true;
-    }
-    return false;
-  }, [nft, isOwner, user.id, isWrongChain]);
-
-  console.log('isUserCanEnterInAuction', isUserCanEnterInAuction);
-
-  const isUserCanEndAuction = React.useMemo(() => {
-    if (nft && !isWrongChain && nft.is_auc_selling && nft.bids.length && isOwner) {
-      if (nft.standart === 'ERC721') return true;
-      if (nft.standart === 'ERC1155' && nft.owner_auction.find((seller) => seller.id === user.id))
-        return true;
-    }
-    return false;
-  }, [nft, isOwner, user.id, isWrongChain]);
-
-  console.log('isUserCanEndAuction', isUserCanEndAuction);
-
-  const isUserCanPutOnSale = React.useMemo(() => {
-    if (nft && !isWrongChain && isOwner && !nft.is_selling && !nft.is_auc_selling) {
-      if (nft.standart === 'ERC721') return true;
-      if (
-        nft.standart === 'ERC1155' &&
-        !nft.sellers.find((seller) => seller.id === user.id) &&
-        !nft.owner_auction.find((seller) => seller.id === user.id)
-      )
-        return true;
-    }
-    return false;
-  }, [nft, isOwner, user.id, isWrongChain]);
-
-  console.log('isUserCanPutOnSale', isUserCanPutOnSale);
-
-  const nftSellingType = React.useMemo(() => {
-    if (nft) {
-      if (nft.is_selling) return 'sell';
-      if (nft.is_auc_selling) return 'auction';
-    }
-    return '';
-  }, [nft]);
-
-  const currentPrice = React.useMemo(() => {
-    if (nft) {
-      if (nft.is_selling) return nft.price;
-      if (nft.is_auc_selling && nft.highest_bid) return nft.highest_bid.amount;
-    }
-    return 0;
-  }, [nft]);
 
   const getItem = React.useCallback(() => {
     storeApi
@@ -244,19 +120,9 @@ const DetailArtwork: FC<Props> = observer(({ className }) => {
           link={mockData.link}
           likeAction={() => alert('like')}
           dotsAction={() => alert('dots')}
-          price={currentPrice}
-          asset={nft?.currency.symbol || ''}
           growth={mockData.growth}
           growthUsd={mockData.growthUsd}
-          author={nft?.creator.name || ''}
-          authorAvatar={nft?.creator.avatar || ''}
-          tags={nft?.tags || []}
-          description={nft?.description || ''}
-          image={nft?.media || ''}
-          isUserCanBuyNft={isUserCanBuyNft}
-          isUserCanEnterInAuction={isUserCanEnterInAuction}
-          isUserCanPutOnSale={isUserCanPutOnSale}
-          type={nftSellingType}
+          nft={nft}
         />
         <div className={styles.chartAndBidders}>
           <div className={styles.chartWrapper}>
@@ -314,6 +180,6 @@ const DetailArtwork: FC<Props> = observer(({ className }) => {
       </div>
     </div>
   );
-});
+};
 
 export default DetailArtwork;
